bug fixes to core flows
