+++ src/testing/index.js
@@ -13,29 +13,38 @@ function addDummyEvent(
 ) {
     const startDate = new Date()
     startDate.setMinutes(startDate.getMinutes() + offset)
-
-//    type = 'meetingEnd'
-    const intention = 'stay awake'
+    const endDate = new Date()
+    endDate.setMinutes(endDate.getMinutes() + offset * 2)
 
     const evt = {
         summary,
         startDate,
+        endDate,
         id,
         location,
         url,
         description,
-        intention,
         type,
     }
 
     console.log('Adding entry to upcomingEvents..')
     add('upcoming', evt)
 }
-
+function uuid() {
+    return ([1e7] + -1e3 + -4e3 + -8e3 + -1e11).replace(/[018]/g, (a) =>
+        (a ^ ((Math.random() * 16) >> (a / 4))).toString(16)
+    )
+}
 module.exports = async function addTestEvents() {
     console.log('In testing mode...')
 
     const waitFor = LOOK_AHEAD_MINUTES - 0.5
-    addDummyEvent('Event', '1xasd72', waitFor, 'meeting room', 'https://google.com')
+    addDummyEvent(
+        'Event',
+        uuid(),
+        waitFor,
+        'meeting room',
+        'https://google.com'
+    )
     // addDummyEvent("Event", "1", 15.25, "meeting room", "https://google.com");
 }
